fix: notification refactor
