+++ app/components/Notification/Notification.module.css
@@ -8,7 +8,7 @@
   background-color: #f1f1f1;
   border-radius: 20px;
   box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-  z-index: 1000;
+  z-index: 4;
   transition: width 0.3s ease;
   max-width: 90%;
   box-sizing: border-box;
